requirements.txt
fastapi
uvicorn
loguru
python-dotenv
pydantic
pandas
yfinance
requests
PyJWT
pyotp
feedparser
pytest
